+++ Pods/SenseKit/Pod/Classes/BLE/SENSleepPill.h
@@ -12,9 +12,15 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
+typedef NS_ENUM(NSUInteger, SENSleepPillAdvertisedVersion) {
+    SENSleepPillAdvertisedVersionUnknown = 0,
+    SENSleepPillAdvertisedVersionOneFive
+};
+
 @interface SENSleepPill : NSObject
 
 @property (nonatomic, strong, readonly) LGPeripheral* peripheral;
+@property (nonatomic, assign, readonly) SENSleepPillAdvertisedVersion version;
 
 - (instancetype)initWithPeripheral:(LGPeripheral*)peripheral;
 - (NSString*)name;