+++ SleepModel/HEMAlarmValueRangePickerView.h
@@ -20,13 +20,13 @@
 
 @property (nonatomic, weak) IBOutlet UIView* topGradientView;
 @property (nonatomic, weak) IBOutlet UIView* botGradientView;
-@property (nonatomic, weak) IBOutlet UIView* separator;
 
 @property (nonatomic, weak) id<HEMAlarmValueRangePickerDelegate> pickerDelegate;
 @property (nonatomic, copy) NSString* unitSymbol;
 @property (nonatomic, assign) NSInteger selectedMinValue;
 @property (nonatomic, assign) NSInteger selectedMaxValue;
 
++ (instancetype)defaultRangePickerView;
 - (void)configureRangeWithMin:(NSInteger)min max:(NSInteger)max;
 - (void)configureWithMin:(NSInteger)min max:(NSInteger)max;