+++ SleepModel/HEMConfigurationsPresenter.m
@@ -117,13 +117,14 @@ static CGFloat const kHEMConfigurationNoConfigSeparatorHeight = 1.0f;
     [skipButton addTarget:self
                    action:@selector(skip)
          forControlEvents:UIControlEventTouchUpInside];
-    [skipButton setHidden:[[self configs] count] == 0];
     [self setSkipButton:skipButton];
 }
 
 - (void)bindWithDoneButton:(HEMActionButton*)doneButton {
     [doneButton addTarget:self action:@selector(save) forControlEvents:UIControlEventTouchUpInside];
     [[doneButton titleLabel] setFont:[UIFont button]];
+    [doneButton setBackgroundColor:[UIColor grey3] forState:UIControlStateDisabled];
+    [doneButton setEnabled:NO];
     [self setSaveButton:doneButton];
 }
 
@@ -192,6 +193,7 @@ static CGFloat const kHEMConfigurationNoConfigSeparatorHeight = 1.0f;
     if ([indexPath row] < [[self configs] count]) {
         config = [self configs][[indexPath row]];
         [self setSelectedConfig:config];
+        [[self saveButton] setEnabled:YES];
         [tableView reloadData];
     }
 }