+++ SleepModel/HEMDeviceService.m
@@ -70,6 +70,10 @@ static CGFloat const HEMPillDfuMinPhoneBattery = 0.2f;
                  object:nil];
 }
 
+- (void)stopScanningForSense {
+    [SENSenseManager stopScan];
+}
+
 - (void)clearDevicesCache {
     [self setDevices:nil];
 }
@@ -145,12 +149,25 @@ static CGFloat const HEMPillDfuMinPhoneBattery = 0.2f;
         || [[self devices] hasPairedSense]);
 }
 
-- (void)findNearestPill:(HEMDevicePillHandler)completion {
+- (void)findNearestPillWithVersion:(SENSleepPillAdvertisedVersion)version
+                        completion:(HEMDevicePillHandler)completion {
+    
     [SENSleepPillManager scanForSleepPills:^(NSArray<SENSleepPill *> *pills, NSError *error) {
         SENSleepPill* pill = nil;
+        
         if (error) {
             [SENAnalytics trackError:error];
-        } else {
+        } else if (version != SENSleepPillAdvertisedVersionUnknown) {
+            for (SENSleepPill* discoveredPill in pills) {
+                BOOL correctVersion = [discoveredPill version] == version;
+                BOOL inDFUMode = [SENSleepPillManager isSleepPillInDFUMode:discoveredPill];
+                if ((correctVersion || inDFUMode)
+                    && [pill rssi] >= HEMPillDfuPillMinimumRSSI) {
+                    pill = discoveredPill;
+                    break;
+                }
+            }
+        } else { // no filter
             // first pill has the strongest signal, but we should only return a
             // pill if it meets minimum RSSI value
             pill = [pills firstObject];
@@ -158,8 +175,13 @@ static CGFloat const HEMPillDfuMinPhoneBattery = 0.2f;
                 pill = nil;
             }
         }
+        
         completion (pill, error);
     }];
+}
+
+- (void)findNearestPill:(HEMDevicePillHandler)completion {
+    [self findNearestPillWithVersion:SENSleepPillAdvertisedVersionUnknown completion:completion];
 }
 
 - (BOOL)isScanningPill {