+++ SleepModel/UIFont+HEMStyle.h
@@ -18,6 +18,7 @@
 + (UIFont*)h4;
 + (UIFont*)h5;
 + (UIFont*)h6;
++ (UIFont*)h6Bold;
 + (UIFont*)h7;
 + (UIFont*)h7Bold;
 + (UIFont*)h8;
@@ -25,6 +26,7 @@
 + (UIFont*)bodyBold;
 + (UIFont*)bodySmall;
 + (UIFont*)bodySmallBold;
++ (UIFont*)buttonLarge;
 + (UIFont*)button;
 + (UIFont*)buttonSmall;